refactor(npu): converge the forked npu_integration onto autobot_shared/npu (#12656)
